[API Compatibility] fix
