+++ paconvert/attribute_mapping.json
@@ -187,6 +187,9 @@
   "torch.float8_e4m3fn": {
     "Matcher": "ChangePrefixMatcher"
   },
+  "torch.float8_e5m2": {
+    "Matcher": "ChangePrefixMatcher"
+  },
   "torch.float8_e4m3fn.itemsize": {
     "Matcher": "ChangePrefixMatcher"
   },
@@ -239,6 +242,15 @@
     "Matcher": "ChangePrefixMatcher"
   },
   "torch.short": {
+    "Matcher": "ChangePrefixMatcher"
+  },
+  "torch.uint16": {
+    "Matcher": "ChangePrefixMatcher"
+  },
+  "torch.uint32": {
+    "Matcher": "ChangePrefixMatcher"
+  },
+  "torch.uint64": {
     "Matcher": "ChangePrefixMatcher"
   },
   "torch.uint8": {
