+++ tests/test_distributions_Distribution_rsample.py
@@ -63,3 +63,15 @@ def test_case_4():
         """
     )
     obj.run(pytorch_code, ["result"], check_value=False)
+
+
+def test_case_5():
+    pytorch_code = textwrap.dedent(
+        """
+        import torch
+        normal_dist = torch.distributions.Normal(torch.tensor([1.0, 2.0]), torch.tensor([0.5, 1.5]))
+        sample_shape = (2, 1)
+        result = normal_dist.rsample(sample_shape)
+        """
+    )
+    obj.run(pytorch_code, ["result"], check_value=False)
