+++ tests/test_nn_Module.py
@@ -82,3 +82,45 @@ def test_case_2():
         """
     )
     obj.run(pytorch_code, ["result"], atol=1e-5, rtol=1e-5)
+
+
+def test_case_3():
+    pytorch_code = textwrap.dedent(
+        """
+        import torch
+
+        class Demo(torch.nn.Module):
+            def __init__(self):
+                super().__init__()
+                self.weight = torch.nn.Parameter(torch.tensor([1.0, 2.0, 3.0]))
+
+            def forward(self, x):
+                return x * self.weight
+
+        x = torch.tensor([0.5, 1.5, 2.5])
+        result = Demo()(x)
+        """
+    )
+    obj.run(pytorch_code, ["result"])
+
+
+def test_case_4():
+    pytorch_code = textwrap.dedent(
+        """
+        import torch
+
+        class Demo(torch.nn.Module):
+            def __init__(self, scale):
+                super().__init__()
+                self.scale = scale
+                self.register_buffer("bias", torch.tensor([1.0, -1.0]))
+
+            def forward(self, x):
+                return x * self.scale + self.bias
+
+        args = (torch.tensor([2.0, 3.0]),)
+        x = torch.tensor([0.25, 0.75])
+        result = Demo(*args)(x)
+        """
+    )
+    obj.run(pytorch_code, ["result"])
